Ignore in-flight ac suggestions on submit.
- See #1237457

If you hit submit while there's an ongoing ajax request for completion
suggestions the timing can be that you get suggestions overlaying your real
results

This adds a simple flag 'off' that is used to track if we should ignore those
in flight suggestions when they come back.

QA:

Compare the results of the test from the bug on comingsoon vs this branch.
When you hit enter quickly after the 'i' in wiki you should not get the
results popup while you do on comingsoon.