Add ability to skip nav on tabview clicks.
The tabview headers are anchor tags. By default, clicking an anchor tag will
cause the url to update. In tests, this causes issue. This adds a flag to the
tabview widget that enables halting the click event so that we can verify all
effects except the url hash change.

This might also prove useful if we need to reuse the tabview widget, yet not
allow for an instance to effect url state.

QA:

Run the tests and yay no hash left once they're complete.