Remove the bws- prefix on charm/bundle tabs.
- They are ids, so we have to watch out for collisions. I don't currently see
any, but if we hit one we might have to migrate the names over. I did not do
this here so that we can keep things backwards compatible with a simple rule
of s/bws-// for the name of the tab.

QA:

Go to a charm and select a tab. The url should be updated to show the #<tab>
and it should not have any bws- in it.

Paste that url into a new window and it should route back correctly for you.

Add a bws- in front of the hash and it should auto remove it and still route
you to the correct tab on the correct charm.