Add in relation menu with ability to delete
Relations now provide a menu with all of the actual relations in the collection
available.  Clicking on the relation will bring up the delete dialog.

Proper positioning of the menu is not included in this branch, but rough
posititioning is.

Switching to quoted keys for window.flags

Lingering environment view assigned to global; adding template.

Responding to reviews; includes drive-by of destroying view in test_environment_view.js