Bug #1269381 - Clarify keyboard shortcut help.
The shortcut labels had to match the lowercase version of the key. In
some cases that didn't make as much sense as the uppercase version. For
example, Shift+/ brings up the help, but Shift+? (the uppercase version)
is more intuitive. Add a way to separate the label displayed from the
actual keybinding (which had to be the lowercase version).