Correct issues around hiding/showing details.
When hiding/showing charm and entity details, there are a host of polish
issues related to the state changes in the il flag. We lost the smooth
animation of the details panel sliding out. The panel contents weren't
being destroyed properly on close. Clicking between tabs would trigger
re-renders and mess up the hash in the URL. All of these issues are now
fixed in this branch.