Subordinate lines update properly
Subordinate relations were not being shown as such in certain instances, such
as when the bundle was deployed through the deployer, since bundles do not
contain subordinate information.  Information about subordinate status was
retrieved once the deploy of the bundle finished and services were associated
with charms.  This adds functionality that informs the whole topology when the
subordinate status of a service changes.