Move AS event handling from browser to view.
Having event handling in the browser made sense when the approach was to
propogate an event over to the topology, but once we switched to the
database as a single source of truth, this code became very clunky.
Moving it into the view allowed for much cleaner method and should make
debugging much easier as the code is spread across fewer files.