Implement the inital cross model relation endpoints.
Implement offer and listOffers, used for offering
endpoints and retrieving a list of remote services.

I presume QA for this would be done in the cmr sprint,
as these API calls should be used to create and list offers.

Note that sandbox mdoe is not implemented yet, and will be
later.