Cross model relations: offer details and remote service info.
Implement the missing ServiceOffers API call on the CrossModelRelations
facade. It is used to retrieve details on a offered remote service.

Also implement the model layer for remote services. It is represented
in the database by a RemoteService model, and its corresponding
model list.

A mega-watcher handler is also introduced in order to populate the
new models. As juju-core does not yet implement relating to remote
services it is not possible to really test the mega-watcher handler
at this time, but I followed to Go code and that logic should be
ready very soon anyway.

Finally, added an addDetails method on the RemoteService that can be
use to enhance the model instance (created on delta) with info
obtained from calling env.getOffer().

Note that sandbox mode is not supported yet.

I expect small tweaks will be required when the functionality is
completed in core, but for now this should be enough for proceeding
with visualization steps.