Improve closing WebSocket connections.
Properly shut down the mega-watcher before quitting
the connection, because that's The_Right_Thing to do and
so that close acks are not deplayed when the GUI is
connected directly to core.

Also better define responsibilities between the base
environemnt and the Go implementation when a connection
needs to be closed.