Properly handle needing to access multiple users.
Used to be that only one username was passed to UserProfile. Turns out
that's not sufficient. We need the "blessed" user (the one returned by
app.js' _getAuth function) for display purposes, but behind the scenes
we also need the charmstore user. We need charmstore to properly
request charms and bundles, and to correctly determine if we need to
show the "Login to charmstore" button.