Prevent deep prop changes from saving instantly.
It seems that when you have a prop that is a complex object, and you
change one of its attributes, both nextProp and this.prop in
componentWillReceiveProps are going to point to the same object.
Consequently comparing them isn't going to work. The solution is to pass
in a clone as the prop value, so the link to the original object is
broken.