Improve login/logout handling.
When logging out now WebSocket connections are properly closed, and the API instances (controller and model) are properly cleaned up in order to avoid any information leak from theoretically logged out connections and from the in-memory database.

Included in this work, a bug prevening login/logout/login again has been fixed by correctly handling the existing credentials when logging out and then logging in again. For instance credentials are now explicitly shared when switching to another model.

This branch also includes extensive sanity improvements to our test suite, especially the old app tests. As part of those improvements, a bug causing the app to dispatch after it's destroyed is also fixed, by cancelling the timer for the method called when the database receives changes.

The _cleanups machinery for the testing suite has been fixed so that at least:
- _cleanups registered functions are called before afterEach (and not after);
- afterEach functions can now use done without the usual "done called more than once" issue (one done was called by the wrapped function, and another from the monkey patched newAfterEach).

Finally, attenuate the horror in the app tests.