Simplify ECS code, and fix pending expose callbacks.
This fixes https://github.com/juju/juju-gui/issues/2297 by addressing the underlaying issue: not providing callbacks to some API calls causes the ECS to not being able to complete.
In that scenario, for instance, expose callbacks are never called because the existing code failed to identify the missing argument and react accordingly.
This branch makes calling the ECS lazy methods explicit, so that we are more in control of what arguments are passed. As a consequence, some magic is also removed, which is always good.

To QA this, deploy the canonical k8s in sandbox mode and on real models. Also please exercise all the API calls and ECS interactons, as this branch introduces major changes in the logic there.

Conflicts:
	jujugui/static/gui/src/test/test_env_api.js
	jujugui/static/gui/src/test/test_env_change_set.js
	jujugui/static/gui/src/test/test_env_legacy_api.js